# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medical Laboratory API",
    "version": "1.0.0",
    "description": "API documentation for Medical Laboratory system (Labs & Subscriptions)"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Subscription": {
        "type": "object",
        "properties": {
          "subscription_id": { "type": "string" },
          "subscription_name": { "type": "string" },
          "subscription_price": { "type": "string" },
          "subscription_duration": { "type": "string", "format": "date-time" }
        }
      },
      "Lab": {
        "type": "object",
        "properties": {
          "lap_id": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" },
          "subscription": { "$ref": "#/components/schemas/Subscription" }
        }
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/auth/lab-login": {
      "post": {
        "summary": "Authenticate a laboratory (Lab Login)",
        "tags": ["Labs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lab login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string" },
                        "labId": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/labs/subscription": {
      "post": {
        "summary": "Register a new laboratory (with subscription)",
        "tags": ["Labs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "subscription_id": { "type": "string" },
                  "subscription": {
                    "type": "object",
                    "properties": {
                      "subscription_name": { "type": "string" },
                      "subscription_price": { "type": "string" },
                      "subscription_duration": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Laboratory created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "lap_id": { "type": "string" },
                        "subscription_id": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Subscription plan is required"
          }
        }
      }
    },
    "/labs": {
      "get": {
        "summary": "Get all laboratories",
        "tags": ["Labs"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Laboratories fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Lab" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/labs/{labId}": {
      "get": {
        "summary": "Get a laboratory by ID",
        "tags": ["Labs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Laboratory fetched successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Lab" }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a laboratory",
        "tags": ["Labs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "subscription_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Laboratory updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete a laboratory",
        "tags": ["Labs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Laboratory deleted successfully" }
        }
      }
    },
    "/subscriptions": {
      "get": {
        "summary": "Get all subscriptions",
        "tags": ["Subscriptions"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Subscription plans fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Subscription" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new subscription plan",
        "tags": ["Subscriptions"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["subscription_name"],
                "properties": {
                  "subscription_name": { "type": "string" },
                  "subscription_price": { "type": "string" },
                  "subscription_duration": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "subscription_id": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/subscriptions/{subscriptionId}": {
      "get": {
        "summary": "Get a subscription plan by ID",
        "tags": ["Subscriptions"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription plan fetched successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Subscription" }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a subscription plan",
        "tags": ["Subscriptions"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subscription_name": { "type": "string" },
                  "subscription_price": { "type": "string" },
                  "subscription_duration": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Subscription plan updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete a subscription plan",
        "tags": ["Subscriptions"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "subscriptionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Subscription plan deleted successfully" }
        }
      }
    }
  },
  "tags": [
    { "name": "Labs", "description": "Laboratory management" },
    { "name": "Subscriptions", "description": "Subscription plan management" }
  ]
}